+++ CHANGELOG.md
@@ -4,5 +4,6 @@
 
 * **setup:** CE-1261 Setup biodiversity analytics project ([CE-1261](https://jira.rfcx.org/browse/CE-1261))
 * **setup:** CE-1297 Setup biodiversity analytics deployment ([CE-1297](https://jira.rfcx.org/browse/CE-1297))
+* **setup:** CE-1335 Setup navigation bar with project selector ([CE-1335](https://jira.rfcx.org/browse/CE-1335))
 
 * **features:** CE-1327 User can see the comparison box at the top of report ([CE-1327](https://jira.rfcx.org/browse/CE-1327))